feat: delete function
feat:
* Delete function - Linked on homepage and on profile page.

fix:
* Set up correct endpoints for fetching on profile page - Previously only supported "own" endpoint.
* Changed button creation on profile page.
* Improved exported values for consistency
* Improved the comparison between inspected profile and logged-in profile.  (now uses simpler logic and exported variables)

+++ src/js/api/post/delete.js
@@ -1,26 +1,37 @@
-import { API_SOCIAL_POSTS } from "../../api/constants";
-
+import { API_SOCIAL_POSTS, API_KEY } from "../../api/constants";
+import { getMyName, getMyToken } from "../../utilities/getInfo.js";
 export async function deletePost(id) {
-    fetch(`${API_SOCIAL_POSTS}/${id}`, {
-        method: "DELETE",
-        headers: {
-          'Content-Type': 'application/json',
-          'X-Noroff-API-Key': API_KEY,
-          'Authorization': 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiQWtzZWxfb2xkZWlkZSIsImVtYWlsIjoiYWtzaGVsODc3MDdAc3R1ZC5ub3JvZmYubm8iLCJpYXQiOjE3MjcxMjYxNjh9.kTNufOOgrial4IJ1MjYPrtdj2ecCzzYRcuyE-vRVnkk'
-        }
-      })
-      .then(response => {
-        if (!response.ok) {
-          throw new Error('Network response was not ok');
-        }
-        return response.json();
-      })
-      .then(data =>{
-        alert(`Post deleted ${data}`)
-      }
+    const token = getMyToken()
+    const confirmed = confirm("Are you sure you want to delete this post?");
     
-      )
-      .catch(error => {
-        console.error('There was a problem with the fetch operation:', error);
-      });
-}
+    if (!confirmed) {
+      return;
+    }
+  
+    fetch(`${API_SOCIAL_POSTS}/${id}`, {
+      method: "DELETE",
+      headers: {
+        'Content-Type': 'application/json',
+        'X-Noroff-API-Key': API_KEY,
+        'Authorization': `Bearer ${token}`}
+    })
+    .then(response => {
+      if (response.status === 204) {
+        alert("Post deleted"); 
+        return; 
+      }
+      
+      if (!response.ok) {
+        throw new Error('Network response was not ok');
+      }
+      
+      return response.json();
+    })
+    .then(data => {
+      location.reload(); // Refresh the page
+    })
+    .catch(error => {
+      console.error('There was a problem with the fetch operation:', error);
+    });
+  }
+
